requirements.txt
numpy==1.23.5
pandas==2.0.3
scipy==1.10.1
plotly==5.12.0
yfinance==0.2.25
pyyaml==6.0